Cleanup obsolete iptables binaries
