--- tools/istio-iptables/.gitignore
@@ -1,1 +1,0 @@
-istio-iptables
