--- tools/istio-iptables/pkg/testing/data/data.go
@@ -1,28 +1,0 @@
-// Copyright Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package data
-
-import (
-	"strconv"
-	"strings"
-)
-
-func NOwnerGroups(n int) string {
-	var values []string
-	for i := 0; i < n; i++ {
-		values = append(values, strconv.Itoa(i))
-	}
-	return strings.Join(values, ",")
-}
